Optimize for performance by minimizing rendering of the drag and drop list
